:bug: [#4923] Ensure layout components are properly processed
Layout components (content, softRequiredErrors, columns, and fieldset) are not present in the
submission data, so they should also not be present in the JSON schema. If any of these components
was present inside an edit grid, an attempt was made to generate a schema for them. This failed
because they do not exist in the component registry.

This fix ensures that the schemas for content and softRequiredErrors components are ``None``, and
the schemas for columns and fieldset components represent their respective children, not the
component itself.

Also added additional logging if JSON-schema generation of a component fails.