:recycle: [#5312] Refactor Generic JSON backend
Part of the schema processing is now already done, so we can avoid doing it twice.

Also fixed the value of the required keyword in ``GenericJSONBackendTests.test_no_file_upload_for_multiple_files_component``
When there is no file upload, we send an empty list, not an empty object within a list, so it doesn't make sense to leave
the required keyword empty.