:sparkles: [#5132] Allow multiple authentication options
To give the Yivi plugin, and end-users, more flexibility we now allow multiple authentication attributes (this config has been renamed to 'authentication options').

This allows end-users to choice themselfs with which authentication option they wish to identify. When only one authentication option is selected, this will become the only/required choice.

When no authentication option is specified, we use a pseudo value to identify the user. This allows the Yivi signature authentication with only e-mail address, or any other value.