:lipstick: [#5132] Allow multiple authentication options in UI
To give the Yivi plugin, and end-users, more flexibility we now allow multiple authentication attributes (which has been renamed to 'authentication options').

In the admin the configuration has been changed to a multiple select input for the authentication options. The admin form changes dynamically based on the selected authentication options.