:sparkles: [#5132] Extend Yivi plugin with bsn, kvk and pseudo configuration
Added configuration fields for full bsn, kvk and pseudo authentication.

All claims in CLAIMS_CONFIGURATION are optional, seeing as we don't know which authentication option ends up being used by an end-user.

Along with basic claim and loa configuration, the Yivi plugin now also contains "Available scopes" configuration. Describes the scopes that are available for the Yivi plugin, and which claims are supposed to derive from it.