:building_construction: [#5140] Authentication options forms using Formik
The authentication options for DigiD have been moved from inline to a modal form. This creates a contained environment for auth plugin configuration. This also reduced the "noice" of the formbuild a bit.

This allows us to use Formik for the auth options forms. Making it easier to support/maintain/develop the authentication configuration.