:sparkles: [#5132] Add yivi scopes to authentication request
Add the Yivi authentication and additional scopes to the login request, using the Signicat format for additional parameters: https://developer.signicat.com/broker/signicat-identity-broker/authentication-providers/yivi.html#example-of-adding-condiscon-parameter-in-your-oidc-request, and the Yivi codiscon format and logic: https://irma.app/docs/condiscon/.

For the login request, the default Yivi scopes are used (defined in the Yivi admin config) in addition to the form/plugin specific defined scopes for authentication and additional information.

When multiple authentication options are allowed by the plugin/form config, the end-user (citizens) can choose which to use for authentication. When only one is defined, that one becomes required. All the "additional_attributes_groups" are optional.