Fixed tests in family members prefill plugin
The tests were not right because they were testing the old functionality (user
should add 2 user defined variables for enabling the family members
plugin, but now the case is different. One user defined variable and the
custom component partners are needed). See the docs for the plugin for
all the instructions.