:bug: [#5041] Ensure fields with a period in their key are added in the Generic JSON plugin
These fields were not added because we iterated over the submitted data, which only contain the top-level keys for these
components. This causes the check whether the key is specified in the options to fail, as these contain the full paths
including periods.

Now, the ``values`` are built up the other way around, which works because ``all_values`` is a ``FormioData`` instance
that allows nested-key access.