:sparkles: [#5132] Yivi anonymous login and additional information hashing
Yivi requires at least one attribute to be provided for authentication to be deemed successful.

By changing the pseudo claim to a Yivi attribute, and by forcing the pseudo claim to be used during anonymous login (i.e. when no authentication options are selected), we can ensure that the Yivi request always works for users without BSN or KVK data.

When hashing the authentication data, we must also hash the additional claims. These could contain personal information, that cannot be exposed.