:green_heart: Fix e2e test on Webkit
In CI, there appears to be some label-overlap check with the checkbox
sudo for the webkit browsers, which supposedly captures
the click.

Verified against the webkit browser 'Browser' on Arch that clicks do
properly propagate. There may be some minor overlap because the label
sits flush against the border of the checkbox, while the checkbox has
a padding of 2px to the right that may trigger this overlap detection.

Either way, the label is associated with the input, so that clicking
the label actually toggles the checkbox anyway.

Forcing the click doesn't seem to help, so as a last resort, the
statement checkboxes are disabled entirely.