:construction: [#5132] WIP for safer attribute groups edit and delete.
The Yivi `FormAuthenticationBackend`'s use `AttributeGroup` for the additional attributes per form. This is saved in the options, which is a json blob.

Because there is no real connection between the `AttributeGroup` and `FormAuthenticationBackend`, editing and deleting of the `AttributeGroup` doesn't reflect to the `FormAuthenticationBackend`.

This branch/change is to solve this issue. Editing the name of the `AttributeGroup` should also change its appearance in any Yivi `FormAuthenticationBackend` `additional_attributes_groups`. Deleting a `AttributeGroup` should not be possible when its used by any Yivi `FormAuthenticationBackend`.