:lock: [#5430] obfuscated Yivi and eIDAS claims for logging
For the claims logging, we need to obfuscate all sensitive Yivi and eIDAS claims.

Yivi:
Because municipalities control which additional attributes are requested during login (which results into the claims), we cannot know which claims are and aren't sensitive. To be safe, we should handle all Yivi additional claims as sensitive. Besides that, we also handle the bsn, kvk and pseudo claims as sensitive.

eIDAS for person:
For personal eIDAS authentication, we regard the legal identifier and additional personal data (firstname and lastname) as sensitive.

eIDAS for companies:
For eIDAS authentication as a company, we regard the legal and acting identifiers as sensitive. In addition, the additional acting personal data (firstname and lastname) are also handled as sensitive.