:recycle: [#2324] Combine methods to get data in the submission value variables state
We always want to use native Python objects, so that will be the standard for `.get_data`.

Also refactored `get_variables_for_context` to use the state for getting data. Note that
there *might* be a difference in what this function outputs, as previously the static data
would be overwritten by the submission data in case of key conflicts. Our validation guards
against the use of duplicate keys, but this does not include old forms or newly-added
static variables.