:bug: [#2324] Fix ZGW API registration not handling native Python objects
Need to add the date formatting to the `rol_mapping`, because the submission data for this
field will be a date object.

To remove the on-the-fly date conversions inside `process_according_to_eigenschap_format`,
we ensure that the values (submission data) passed to this routine are fetched from the
state properly.