:recycle: [#2324] Add data type fields to submission value variable
To ensure data type information is not lost on deletion of forms/form steps/form variables,
we save the data types to the SubmissionValueVariable as well.

Note that this may lead to inconsistencies if a form/form step/form variable was changed,
as we no longer get the data type from the form variable. The general idea is to take a
"snapshot" of the complete form configuration and attaching it to the submission. This
means that changes in the form will no longer affect existing submissions of that form.
Though the scope of this change is too big to tackle here, so we make a small start with
the data types first.