:bug: [#2324] Ensure partners data is converted to native Python objects
The partners data contains a date of birth, which is rendered and formatted as a date
component, and since the formatters do not support strings anymore, we need to make sure
it is converted to a date object.

Note that using the form variable data (sub)type is a work-around, and might lead to
problems in the future. For now, we always expect the ``dateOfBrithPrecision`` to be
'date' (which is also what ``PartnersSerializer`` currently does). If this changes,
partial dates (month-year, year, etc.) might be interpreted as a full date (ISO) string,
and converting to Python types will result in ``None``.