:heavy_plus_sign: [#5451] Add dependencies to export telemetry via the OTLP protocol
These libraries are necessary to be able to export to the OTel collector
running in docker compose.

Additionally, many observability sinks (like Sentry, APM, Prometheus...)
support the otlp protocol directly, so most likely we'll be able to
get by with this single dependency. Otherwise the recommendation will
probably be to set up a collector and configure exporters appropriately.