:sparkles: [#5454] Allow configuring the Cross-Origin-Opener-Policy
Piwik Pro debug mode requires this header to be relaxed, but we'd prefer
to use sane/secure defaults. The setting can now be configured through
an environment variable for those that wish to use it.

This header is currently *NOT* part of the headers checked/enforced by
the DigiD audits, but it's reasonable to expect it will be in the future
when security is tightened. When that happens, our policy will be to use
a browser extension and we'll drop support for configuring it via envvar
or it's up to the municipality to explain in their audit why it's set
to an unsafe value.