:card_file_box: [#2324] Add component configuration to submission value variable
To be able to convert submission data for editgrid components to native Python types,
we need the component configuration, to fetch data type information of their children.

We can get this through submission -> form -> form step -> form definition, but it
will not be available for submissions that have their form/step/definition deleted.
Therefore, the component configuration is added to the submission value variable.

Also includes a script to set the configuration to existing submission value variables.
If a related form/step/definition was deleted, we can get the configuration from the
form step history on the submission step. A migration to perform this operation was
considered, but turned out to be not viable for large production environments (10M
variables). It would take over an hour to complete...