:sparkles: [#5475] Additional claims accessible in jsonlogic and as template variables
The `additional_claims` of authInfo could be problematic for jsonlogic and the django templating. These two ways of using variables don't support periods and hyphens ("." and "-") in variable keys. This is mainly the case with Yivi attributes.

To make all variables accessible for jsonlogic and django templating, we replace the periods and hyphens with underscores.