:chart_with_upwards_trend: [#5451] Enable open telemetry integration
In CI and dev mode, the Open Telemetry SDK is disabled by default. For
other environments, output will be written to stdout/stderr if the
OTLP endpoint cannot be reached, but it will not interfere with the
functionality of Open Forms. Come near release time, we'll have to
decide if we make this opt-in or opt-out.

Each logical unit init script has been provided with the service
name environment variable so we can properly identify them. Local
development workflows fall back to a default, though you can also
set a value in .env.