:whale: [#5451] Add OTel collector and prometheus to observability stack
The observability stack now includes the open telemetry collector for
local testing, and uses prometheus to store metrics so that the full
observability toolchain can be tested and used locally.

It can be used as an example to set up your own infrastructure, but
we don't claim that this is *the* stack.