:wastebasket: [#5451] Deprecate the elastic apm library usage
Elastic APM mostly captures spans and is definitely useful, however,
it's a vendor-specific solution and the Open Telemetry SDK can capture
spans too - and more.

A more vendor-agnostic setup is instrumenting the application with OTel
and sending the spans to the collector (which happens as of this
ticket) and the export from the collector to Elastic APM (or Jaeger,
or whatever backend you want...).

We start deprecating now, so that we can properly deprecate it once
infrastructure has been migrated and we perhaps have some metrics about
how many organisations actually uses Elastic APM, before deciding to
remove the library/dependency.