:card_file_box: [#2324] Add migration for data (sub)type fields
We need to fill these fields for existing form and submission variables.
Updating form variables in a migration will *probably* be fine performance wise.

For the submission variables, we can use the existing script that adds the
configuration. Also removed an outdated check on whether the variable key is
in the `components_in_step` dictionary, as we now filter on this in the query
already