:card_file_box: [#2324] Add database constraints to form and submission variables
We can add a constraint on the form variable to ensure that the data subtype is set
when a component has data type array. Furthermore, we ensure that the data subtype
exclusives are not set to the data type.

For the submission variables, we can only check the latter, as we have no information
about the source (component or user defined).