:recycle: Annotate submissions with their stage and refactor data removal functions
Once we start using these stages in multiple places,
it's better to put it in a queryset method to get
consistent usage out of it, which means refactoring
the data_removal tasks.

There was quite some duplication there that can be
abstracted/unified and this was an opportunity to
address the type checker errors too.