:recycle: [#5134] Fix clear-on-hide behaviour
The previous approach was to handle the clear-on-hide behaviour after backend logic evaluation.
This could result in logic rules being triggered when they shouldn't, as the value of the
trigger component should already have been cleared. Furthermore, it could result in a
difference in data between backend and frontend, as simple conditional logic is already
evaluated by the frontend.

This solution solves both problems, by splitting the handling of clear-on-hide behaviour in
two parts:
  1. Process conditionally and statically hidden components before evaluating backend logic.
     Note: statically hidden components are components that have no conditional or backend logic
     applied to them.
  2. Move the handling of clear-on-hide behaviour to the property action, which means mutations
     are applied to the data directly.

Part 1 ensures we are not operating on (possibly) outdated data, as it *should* now match the
state of the frontend. "should", because at the time of writing this we are still dealing with
Formio... This change is meant to go together with the implementation of our own renderer.

Part 2 ensures that logic rules that are dependent on the result of previous actions can operate
with data that is accurate. Note that moving it to the property action only covers the case when
a component goes from visible to hidden when the logic rule is triggered. If a component is
hidden by default, and a logic action that sets it to visible is NOT triggered, we also need to
clear it. This is why we also handle the clear-on-hide behaviour separately in
`iter_evaluate_rules` using `handle_clear_on_hide`. See the inline comments for more information.