:chart_with_upwards_trend: [#3999] Add metric to report component type usage per type and form
For now, we opt to let this metric/callback execute on the metric export
interval (every 60s) by default. In theory, depending on how many forms
and how complex they are, this could take longer-than-desired processing
time, but:

* metric exporting runs in a separate thread
* the queries are not that complicated and limited in amount

If a need arises to optimize this, we can switch to a generator approach
and keep track of the last 'scrape' so that we only scrape every hour or
10 minutes.

Another approach could be to run a single global scope metric exporter
container whose sole purpose is scraping the database - this could be
exposed through a management command and configuration with envvar, but
it increases operational complexity.