:recycle: [#5419] Replace Yivi attribute periods with underscores during claim processing
The Yivi attributes use periods in the attribute names, this is problematic for jsonLogic, django templating and for glom.

Previously, in https://github.com/open-formulieren/open-forms/issues/5475, we solved this by replacing the periods when using the static variables.

Now with the Yivi prefill plugin this same issue arises. So solve this once and for all, the Yivi attributes will be cleaned during the processing of the claims.