:recycle: [#5479] Refactor eIDAS for Companies claim fields
Instead of using a generic identifier field for the acting subject, which would allow bsn, other national identications and pseudo values, we now support specific bsn and pseudo claims.

To simplify the authentication for eIDAS for companies, and to properly identify the used authentication claims, we replaced the generic acting subject identifier fields with specific identification claim fields.