:card_file_box: [#5479] Migrating authInfo and registratorInfo attribute values.
We no-longer support the value `national_id` for AuthInfo and RegistratorInfo `attribute`, as we cannot reliably identify national identifying values besides BSN. This option was added for the eIDAS via OIDC and eIDAS for companies via OIDC authentication plugins.

When logging in via eIDAS via OIDC and eIDAS for companies via OIDC using a national identification value (which isn't BSN), the `attribute` should be set to `pseudo`.

Besides not being able to reliably identify non-BSN identity values, we currently also don't perform special actions that rely on this value.