:safety_vest: [#5479] Validate eIDAS over OIDC and eIDAS for companies over OIDC processed claims
To identify the authenticated user, or the acting subject (in case of eIDAS for companies), we require an identifying value. This could either be BSN or PseudoID (for international identifying values).

Using the `ClaimProcessingInstructions` we cannot expect that either BSN or PseudoID is present in the claims. So, we mark them both as optional, and validate the claims after processing.

Using `validate_processed_claims` we check if either of the identifying values is present. If not, we raise an `ValueError`, which is picked-up in `get_or_create_user` in the `BaseDigiDeHerkenningPlugin` definition.