:children_crossing: [#5515] Updated import handling of Yivi authBackends
The import handling of Yivi authBackend `additional_attributes_groups` will try to update old attributeGroup references to the new notation (using UUID's instead of group names as value). If we cannot correct the import data, we just pass the original data along.

We unfortunately cannot simplify the `SlugRelatedField` queryset. If we change the queryset to `queryset=AttributeGroup.objects.all()`, it expects AttributeGroup objects as input. When working with this field, or exporting the data, this is fine. But during import, the exported data is just a list of UUID's, which raises exceptions as it isn't conform the expected format.