:sparkles: [#5419] Add endpoint for Yivi prefill attributes
Expose the Yivi prefill attributes using a custom endpoint, as the Yivi prefill attributes also contain information about when the prefill attributes can be chosen by the admin.

To dynamically configure the Yivi prefill plugin, we decide the available prefill attributes in the frontend.

There are two reasons why a Yivi attribute would be shown:
- Either its a static prefill attribute, like for retrieving auth identifier data.
- Or the related attributeGroup is configured on the Yivi authentication plugin (in which case, the attributes might be resolved after authentication and prefill-able)