:white_check_mark: [#5134] Update tests
Most of these tests were failing because we now check for components in the registry during
logic evaluation, and crash hard if it could not be found. This means typos in tests (and in
general) are deadly :skull:

`test_report_is_generated_in_same_language_as_submission`:
Removed some outdated fields from this test, and changed how the editgrid is generated

`test_components_hidden_by_frontend_after_filling_have_correct_empty_value` and
`test_components_hidden_by_frontend_have_correct_empty_value`:
Moved these tests to `ConditionalLogicTests`, because it is difficult to check these things through
the endpoint. This is because we no longer return the changes due to conditional logic, but just
save it to the state without persisting to the database.