:bug: [#5557] replace regex pattern for filenames since it's not used for 'originalName' in formio with just soft hyphen removal
:recycle: [#5557] sanitize "originalName" attribute

Formio doesn't change the names of uploaded files. Django on the other hand does change them.
When the file is being uploaded "temporary-file-upload" endpoint is used, which has
MultiPartParser, which changes the file name in the 'sanitize_file_name' method,
including the removal of soft-hyphens.
Formio isn't aware of change, so its "originalName" attribute may differ from the file name
stored in django. Here we apply the same changes to the "originalName" attribute

Backport-of: #5560