:whale: [#5513] Set up sample OTel instrumented nginx image/container
Enabling the otel module in nginx (docker compose) when the
observability stack is not up will print some errors/warnings,
but not block actual functionality, i.e. requests will still
be properly served. It just can't export any telemetry data.

It does seem that the OTel exporter configuration applies for the
entire server - which shouldn't be much of a problem, but it
does show that there's limited flexibility - in particular
setting the otel_service_name may be a problem if fine
grained traces are desired. OTOH you'll probably want to filter
by trace attributes like host name/path so it's a bit
unclear what's the best practice here.