:bricks: [#5060] Add sample redis sentinel setup
This setup is only for illustration and development purposes. It is not
a reference on how to deploy redis sentinel - for that you should use
proper infrastructure (read: use an operator on Kubernetes).

This setup validates that a master with one replica can be deployed,
which is monitored by three sentinel instances.

The addresses of the sentinels must then be configured in the Celery
broker configuration so that it gets information about the (current)
master when a failover happens.

Operating Redis and Redis Sentinel is really outside of the scope of
the Open Forms project as we don't have hands-on experience with such
setups.

Note that the sample docker compose setup uses host networking mode
for the proper mechanisms the work, as the NAT performed in docker
bridge networking is a known limitation.