[#4879] save the payment ID in the provider_payment_id field
Previously the merchant reference was saved in that field but this was not correct. The payment ID allows us to perform API requests to the Worldline API whereas the merchant reference does not. The merchant reference is still saved though in OF through the public_order_id field (which is generated by OF).