:building_construction: [#5419] Ensure new auth plugin hook in prefill plugin is called
A new hook was added to check that a plugin can run only
if a particular auth plugin was used (required for eIDAS
because the auth_attribute check is not granular enough).

The first iteration would call this hook in the plugin itself,
but rather this should be called by the generic code and skip
plugin evaluation entirely if the conditions are not met.

This also made me realize that the options-based plugin
evaluation doesn't check whether the plugin is enabled or
not, which has now been added too, and the type checker
errors have been resolved + the module added to the list
of modules to type-check in CI.