:recycle: [#5419] Ensure the same prefill plugin warnings are displayed in various contexts
The prefill configuration from components was already being checked and
displayed on the form definition panel, but for similiar prefill config
on user defined variables, no such warnings are displayed.

This is even more relevant for the eidas prefill options, which allows
us to now use the generic/default edit modal to select a plugin +
attribute, removing the need for the custom configuration forms
with their own warnings.