:children_crossing: [#5575] Add email digest check for map component overlays
Because the configuration of map component overlays is
rather loose, we should periodically check if the
configuration is still correct.
A number of problems could arise, such as: WMS
resources that change their available layers,
WMS tile layers that are deleted in OF,
WMS resource urls that are no-longer available.

In all these situations, the form admin should
be notified about problems in their map components.