:white_check_mark: [#5419] Update test for expected Yivi attribute transformations
The intent is to destructure the Yivi attributes by splitting on the
period character to create a nested structure, so that we will
actually be able to process this data structure with JSON logic
without needing to mangle the periods into underscores.

That should also naturally make things properly available in the DTL,
removing the need for a couple of workarounds.