:art: [#5419] Destructure Yivi attributes
Instead of replacing '.' characters with an underscore ('_') in the
Yivi attributes and requiring workarounds in form logic/templates,
we now split on the characters and created nested objects in the
returned attributes from the claims.

This is a more natural approach and the test proves that deep merges
happen without overwriting a parent scope.