:sparkles: [#5419] Add initial Yivi prefill plugin foundation
The Yivi prefill depends on the selected attribute groups in the
Yivi authentication plugin configuration, so we cannot statically
return a list of available attributes.

Additionally, there are some possible fixed attributes - in the auth
plugin options you can specify if you want bsn, kvk and/or pseudo or
a subset of these. They all get saved in the same location eventually
since they're mutually exclusive, but you do want to be able to prefill
that value as well.

We can leverage the custom plugin configuration context which gets
serialized from the plugin API endpoint to provide the possible fixed
attributes to the frontend code. The mechanism here is that we expose
each possible auth attribute that can be selected in the authenticatoin
plugin options as an individual attribute. If you provide KVK while
BSN is set for prefill, then nothing will happen, i.e. it only kicks in
if the auth attribute spec matches and then the value will be prefilled.

If the particular auth attribute is not relevant, we also expose the
identifier value itself which can be one of either, in case that needs
to be prefilled too. That will always resolve.

Additional important details:

* requires_auth_plugin is specified so that we can show warnings if the
  yivi auth plugin is missing (meaning no prefill can run, and this is
  checked by the low-level code)
* for_components is set to an empty collection, which effectively
  prevents the plugin from being listed in the form-builder modal (
  prefill tab)