:zap: [#4951] Cache generating a map image
The map is generated from images fetched from an external service. To avoid fetching
these over and over for the same parameters, we can cache it.

Note that the cache is kept relatively small, because we expect only a few calls
within a short time span: generating a single submission PDF and/or confirmation email.
I think it's safe to assume the drawn shapes on the map across different submissions
are different