:recycle: [#5419] Move shared Yivi functionality into shared location
The attribute groups are read in both the authentication and prefill
plugin context, which requires the code to be organized so that it's
more clear these fetchers and mocks are a shared utility.

Semantically, the endpoint is still owned by the authentication layer,
as prefill cannot work without it, but the other way around is
possible.