:poop: [#5419] Found a way to make the prefill tab work
By hijacking the callback to produce the attributes for a given
plugin and targetting the yivi identifier, we can actually make the
dynamic mechanism for attribute derivation kick in.

This does require some pretty nasty workarounds by consuming the react
context in the formio builder wrapper and setting that as custom
options in the web form builder class/instance of Formio.js so that we
can include it in the closure passed to our formio-builder props/context.