:bug: [#5464] Properly deserialize backend options first
Ensure that the plugin configuration options serializer is
used to deserialize the database JSON field options before
passing them into the plugin to test for json-schema
support.

The serializer is a crucial step to handle backwards
compatibility and input sanitation so that plugins can
reason in their rich python types for the options,
as it typically sets some defaults for optional
configuration aspects.