:test_tube: [#5464] Add regression test for KeyError
Despite the field not being required in the options, it
can apparently still cause a crash during JSON schema
generation, which is likely what we saw in Sentry.

Unfortuntely, the event in Sentry is that old that
it has been removed, so we no longer have traceback
or other error information.