:zap: [#4951] Fetch map tiles in parallel
This decreases the time necessary to generate the map image.

Some results with 16 workers:

For a small image map (4 tiles - this corresponds to the 400 by 300 px image size we currently use):
- Sequential: 0.36 s
- Parallel: 0.20 s

For a slightly bigger map (16 tiles):
- Sequential: 1.2 s
- Parallel: 0.40 s

For large maps (100 tiles):
- Sequential: 8.0 s
- Parallel: 0.79 s

Because we usually only need to fetch 4 tiles, hardcoding to 4 workers is sufficient for now.
Even for large maps loading would likely improve from 8 sec to 2-3 sec