[#5622] Added validation for the identifier_variable and removed source_component_type
The validation makes sure that the identifier is part of the data
mappings. The source_component_type was initially added to avoid
searching for the type of the compoenent in the backend since it was
already known on the frontend. This is not robust so we do the check in
the backend.