:alien: Handle Signicat's session stickiness
What used to be a guaranteed new login screen now apparently remembers
an earlier login (with the lower LOA) and the response is not an HTML
page, but a redirect back to the ACS URL, which breaks because
https://localhost:8000 is not actually available (we do some manual
processing to drop the netloc for our test client).

Using different session instances uses different cookie jars, causing
Signicat to see different login sessions and the original test
behaviour is restored this way.