:green_heart: [#5619] Make tile download order deterministic
When multiple tiles are being downloaded, the thread pool
makes the download order non-deterministic as it's IO-bound.

By pinning the pool size on one during tests, we can force the
downloads to happen sequentially, while keeping the optimized
behaviour in production with a larger pool. Because only one
thread is available, the downloads will be performed in the
order that the tasks are scheduled, making them deterministic.