:art: [#5646] Clean up partners/children component nodes
Emit the same layout modifiers like editgrid so that we get the
same styling when generating summaries/PDF.

The custom component with key/type fixed to component_label is not
necessary, because the custom group node (child/partner) controls the
label that's emitted, which is also picked up in the frontend (SDK)
summary page.

This allows us to completely remove the children/partners modifiers for
the styling and re-use the edit grid styling to get the grouping.

Note that the backend code is prepared for a custom group item label,
but this is not yet possible to configure anywhere, so we fall back
to the singular literals 'Partner' and 'Child'.

Note also that generating the (selected) child nodes requires some
extra attention - we need to track the proper index of the selected
value, but possibly keep a different index for the summary so that
those numbers properly increment without gaps while displaying the
correct information.