:bug: [#5134] Fix conversion of empty values for date-related components
An empty value for the date, time, and datetime components is an empty string in Formio, and
`None` in our Python type domain. So we need to make sure it is properly serialized in the
submission step serializer response, and saving to the database.

Note that once we have implemented our renderer, we can revert this, because we will be able
to control the data ourselves

Backport of: #5665

(cherry picked from commit df081a144a053dd20697e1d43478d13ada6105ef)