:sparkles: [#5576] Add map WFS tile layers to admin
WFS tile layers are similar to WMS tile layers, in the way that they both allow additional information to be presented in a map component. WFS takes it one step further, to allow the user to interact with the added visuals.

The WFS version of BAG has been added as default WFS tile layer, to quickly let users become familiar with the new functionality.

The WFS tile layers can be imported and exported from the WFS tile layers configuration page in the admin. This allows quick and easy sharing of WFS tile layers between OF instances.