:bug: Add data subtype to form variable attributes constant
The data subtype is also derived from the matching component, so we need to add it.

When changing the "multiple" property of a component and saving the form (variable), an
integrity error would be raised for the "form_variable_subtype_empty_iff_data_type_is_not_array"
constraint, because this field was not updated.

Backport of: #5667