:lipstick: [#5429] Wrap label text in component preview in modal
The previous fix was for the columns preview in the formio-builder
which doesn't properly vertically align if some labels wrap and others
don't.

This fix is kept for the full form preview, but inside the component
settings modal we can apply the regular label wrapping behaviour as
the columns situation there doesn't apply.