:zap: [#5134] Optimize getting evaluation data for editgrids
Instead of performing a deep copy of the entire data for every editgrid item, we can
do this just once before processing the items. This is because fields inside an editgrid
item cannot affect fields outside the editgrid, so the context outside the editgrid
stays the same. The only thing that is changing is the editgrid data itself, because we
set the item data directly to it on each iteration.