:alien: [open-formulieren/open-forms-sdk#445] Update static asset paths
Starting with v3.4.0 (currently in development), the layout of the
static assets on disk is changing. The ESM and UMD bundles are
emitted in a subdirectory 'bundles', while the ESM modules (for NPM
package users) are now in the top-level directory.

The SDK docker image contains symlinks to mimick the old situation,
but we can already incorporate this properly in our own processing.