:card_file_box: Add default data subtype if no corresponding component was found
If a form variable was changed manually, there is a possibility that we cannot find the
corresponding component definition in the formio configuration. To avoid violating the
data type constraint introduced in
0111_formvariable_form_variable_subtype_empty_iff_data_type_is_not_array_and_more,
we default the data subtype to a string in these cases.

Backport of: #5689

(cherry picked from commit 58f72a813b920c5928d08ae6a379595088215d93)