:card_file_box: [#5722] Update DB models for multiple webhook configuration instances
Multiple PSPIDs on the Worldline side come with distinct
sets of credentials (webhook key ID + secret) that must
each be known in Open Forms. Therefore, the model cannot
be a singleton, but must accept multiple records.

The PSPID field is added - similar to the stable/3.2.x
commit 8bedee7e477c4d2d4830fa523b781e3adde2d472 for easy
identification, but it shouldn't be necessary at runtime.

The actual data migration is moved into a later migration
which the next commit shall update for the updated
db structure/model layout.

Instances that were already on 3.3.0 should not be affected
by this - the change is backwards compatible and the new/
updated data migration will result in a noop.

(cherry picked from commit cf7e197e4c699eb462510e26fef9d0e72716d256)

Backport-of: #5726